add LittleFS
