+++ Smart-LED/data/index.html
@@ -1,0 +1,43 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <link rel="stylesheet" href="style.css">
+    <title>Smart LED</title>
+</head>
+<body>
+    <div class="section-1">
+        <div class="section-block">
+            <h1 class="text">Smart LED</h1>
+        </div>
+        <div class="section-block">
+            <button class="button-control" id="button" onclick="setButtonState()">OFF</button>
+        </div>
+    </div>
+
+    <div class="section-control">
+        <div class="box">
+            <div class="control">
+                <input class="slider red" type="range" id="volume-red" min="1" max="255" value="150" onchange="setLedColor()">
+            </div>
+        </div>
+
+        <div class="box">
+            <div class="control">
+                <input class="slider green" type="range" id="volume-green" min="1" max="255" value="150" onchange="setLedColor()">
+            </div>
+        </div>
+
+        <div class="box">
+            <div class="control">
+                <input class="slider blue" type="range" id="volume-blue" min="1" max="255" value="150" onchange="setLedColor()">
+            </div>
+        </div>
+    </div>
+
+</body>
+
+<script src="script.js"></script>
+
+</html>
