+++ Smart-LED/data/style.css
@@ -1,0 +1,98 @@
+.body{
+    overscroll-behavior: none;
+}
+
+.section-1 {
+    display: flex;
+    justify-content: space-between;
+    width: 100%;
+    height: 15vh;
+    background-color: rgb(38, 38, 38);
+}
+
+.section-block {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    width: 45%;
+    height: 100%;
+}
+
+.text {
+    font-family: 'Franklin Gothic Medium', 'Arial Narrow', Arial, sans-serif;
+    color: #ffffff;
+}
+
+.button-control {
+    width: 60%;
+    height: 50%;
+    border: 2px;
+    border-radius: 50px;
+    background-color: #ba3c3c;
+    font-size: 20px;
+    transition: linear 0.3s;
+}
+
+.button-control:active{
+    transform: scale(0.9);
+    background-color: #8adda8;
+}
+
+.section-control {
+    display: flex;
+    justify-content: space-between;
+    width: 100%;
+    height: 70vh;
+}
+
+.box {
+    display: flex;
+    justify-content: center;
+    width: 30%;
+    height: 95%;
+}
+
+.control {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    flex-grow: 1;
+    width: 50%;
+    height: 90%;
+}
+
+.slider {
+    -webkit-appearance: none;
+    appearance: none;
+    position: absolute;
+    width: 100%;
+    height: 20px;
+    border-radius: 4px;
+    outline: none;
+    opacity: 0.8;
+    transition: opacity 0.3s;
+    transform: rotate(-90deg);
+}
+
+.slider.red {
+    background: linear-gradient(to right, #ffffff, #ba3c3c);
+}
+
+.slider.green {
+    background: linear-gradient(to right, #ffffff, #4fba3c);
+}
+
+.slider.blue {
+    background: linear-gradient(to right, #ffffff,rgb(0, 64, 255));
+}
+
+.slider::-webkit-slider-thumb {
+    -webkit-appearance: none;
+    appearance: none;
+    width: 30px;
+    height: 50px;
+    border-radius: 20px;
+    background: rgb(138, 138, 138);
+    cursor: pointer;
+    transition: 0.2s;
+}
