+++ Smart-LED/platformio.ini
@@ -15,8 +15,8 @@ framework = arduino
 upload_speed = 115200
 monitor_speed = 115200
 upload_port = COM6
+board_build.filesystem = littlefs
 lib_deps = 
-	fastled/FastLED@^3.9.16
 	esp32async/ESPAsyncTCP@^2.0.0
 	esp32async/ESPAsyncWebServer@^3.7.7
 	WiFi.h
